Rename ArrayAny rule
