+++ config/set/php84.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 use Rector\Config\RectorConfig;
 use Rector\Php84\Rector\Class_\DeprecatedAnnotationToDeprecatedAttributeRector;
 use Rector\Php84\Rector\Foreach_\ForeachToArrayAllRector;
-use Rector\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector;
+use Rector\Php84\Rector\Foreach_\ForeachToArrayAnyRector;
 use Rector\Php84\Rector\Foreach_\ForeachToArrayFindKeyRector;
 use Rector\Php84\Rector\Foreach_\ForeachToArrayFindRector;
 use Rector\Php84\Rector\FuncCall\AddEscapeArgumentRector;
@@ -24,7 +24,7 @@ return static function (RectorConfig $rectorConfig): void {
             ForeachToArrayFindRector::class,
             ForeachToArrayFindKeyRector::class,
             ForeachToArrayAllRector::class,
-            ForeachToArrayAnyCombinedRector::class,
+            ForeachToArrayAnyRector::class,
         ]
     );
 };
