--- rules-tests/Php84/Rector/Foreach_/ForeachToArrayAnyCombinedRector/Fixture/boolean_assignment_skip_no_break.php.inc
@@ -1,17 +1,0 @@
-<?php
-
-namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector\Fixture;
-
-class SkipNoBreak
-{
-    public function run(array $animals)
-    {
-        $found = false;
-        foreach ($animals as $animal) {
-            if (str_starts_with($animal, 'c')) {
-                $found = true;
-            }
-        }
-        return $found;
-    }
-}
