--- rules-tests/Php84/Rector/Foreach_/ForeachToArrayAnyCombinedRector/config/configured_rule.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector;
-
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rule(ForeachToArrayAnyCombinedRector::class);
-};
