+++ rules-tests/Php84/Rector/Foreach_/ForeachToArrayAnyRector/Fixture/boolean_assignment_skip_multiple_statements.php.inc
@@ -1,23 +1,19 @@
 <?php
 
-namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector\Fixture;
+namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyRector\Fixture;
 
-class SkipVariableReUseAfterForeach
+class SkipMultipleStatements
 {
-    public function checkAnimal(array $animals)
+    public function run(array $animals)
     {
         $found = false;
         foreach ($animals as $animal) {
             if (str_starts_with($animal, 'c')) {
                 $found = true;
+                echo 'Found: ' . $animal;
                 break;
             }
         }
-
-        if (isset($animal)) {
-            echo 'hit';
-        }
-
         return $found;
     }
 }
