+++ rules-tests/Php84/Rector/Foreach_/ForeachToArrayAnyRector/Fixture/boolean_assignment_skip_non_array.php.inc
@@ -1,16 +1,18 @@
 <?php
 
-namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector\Fixture;
+namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyRector\Fixture;
 
 class SkipNonArray
 {
     public function checkAnimal(\ArrayIterator $animals)
     {
+        $found = false;
         foreach ($animals as $animal) {
             if (str_starts_with($animal, 'c')) {
-                return true;
+                $found = true;
+                break;
             }
         }
-        return false;
+        return $found;
     }
 }
