+++ rules-tests/Php84/Rector/Foreach_/ForeachToArrayAnyRector/Fixture/boolean_assignment_with_key.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector\Fixture;
+namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyRector\Fixture;
 
 class WithKey
 {
@@ -21,7 +21,7 @@ class WithKey
 -----
 <?php
 
-namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector\Fixture;
+namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyRector\Fixture;
 
 class WithKey
 {
