+++ rules-tests/Php84/Rector/Foreach_/ForeachToArrayAnyRector/Fixture/early_return_skip_missing_return_false.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector\Fixture;
+namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyRector\Fixture;
 
 class SkipMultipleStatements
 {
@@ -8,10 +8,8 @@ class SkipMultipleStatements
     {
         foreach ($animals as $animal) {
             if (str_starts_with($animal, 'c')) {
-                echo 'Found: ' . $animal;
                 return true;
             }
         }
-        return false;
     }
 }
