+++ rules-tests/Php84/Rector/Foreach_/ForeachToArrayAnyRector/Fixture/early_return_skip_multiple_statements_after_foreach.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector\Fixture;
+namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyRector\Fixture;
 
 class SkipMultipleStatements
 {
@@ -11,5 +11,7 @@ class SkipMultipleStatements
                 return true;
             }
         }
+        echo 'ok';
+        return false;
     }
 }
