+++ rules-tests/Php84/Rector/Foreach_/ForeachToArrayAnyRector/Fixture/early_return_skip_multiple_statements_within_if_statement.php.inc
@@ -1,19 +1,17 @@
 <?php
 
-namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector\Fixture;
+namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyRector\Fixture;
 
 class SkipMultipleStatements
 {
     public function run(array $animals)
     {
-        $found = false;
         foreach ($animals as $animal) {
             if (str_starts_with($animal, 'c')) {
-                $found = true;
                 echo 'Found: ' . $animal;
-                break;
+                return true;
             }
         }
-        return $found;
+        return false;
     }
 }
