+++ rules-tests/Php84/Rector/Foreach_/ForeachToArrayAnyRector/Fixture/early_return_skip_non_array.php.inc
@@ -1,0 +1,16 @@
+<?php
+
+namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyRector\Fixture;
+
+class SkipNonArray
+{
+    public function checkAnimal(\ArrayIterator $animals)
+    {
+        foreach ($animals as $animal) {
+            if (str_starts_with($animal, 'c')) {
+                return true;
+            }
+        }
+        return false;
+    }
+}
