+++ rules-tests/Php84/Rector/Foreach_/ForeachToArrayAnyRector/Fixture/early_return_skip_unexpected_early_return_other_than_true.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector\Fixture;
+namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyRector\Fixture;
 
 class SkipMultipleStatements
 {
@@ -8,10 +8,9 @@ class SkipMultipleStatements
     {
         foreach ($animals as $animal) {
             if (str_starts_with($animal, 'c')) {
-                return true;
+                return 'ok';
             }
         }
-        echo 'ok';
         return false;
     }
 }
