+++ rules-tests/Php84/Rector/Foreach_/ForeachToArrayAnyRector/Fixture/early_return_with_key.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector\Fixture;
+namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyRector\Fixture;
 
 class WithKey
 {
@@ -19,7 +19,7 @@ class WithKey
 -----
 <?php
 
-namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector\Fixture;
+namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyRector\Fixture;
 
 class WithKey
 {
