+++ rules-tests/Php84/Rector/Foreach_/ForeachToArrayAnyRector/ForeachToArrayAnyRectorTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyCombinedRector;
+namespace Rector\Tests\Php84\Rector\Foreach_\ForeachToArrayAnyRector;
 
 use Iterator;
 use PHPUnit\Framework\Attributes\DataProvider;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
-final class ForeachToArrayAnyCombinedRectorTest extends AbstractRectorTestCase
+final class ForeachToArrayAnyRectorTest extends AbstractRectorTestCase
 {
     #[DataProvider('provideData')]
     public function test(string $filePath): void
