+++ rules-tests/Php84/Rector/Foreach_/ForeachToArrayAnyRector/config/configured_rule.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\Config\RectorConfig;
+use Rector\Php84\Rector\Foreach_\ForeachToArrayAnyRector;
+
+return static function (RectorConfig $rectorConfig): void {
+    $rectorConfig->rule(ForeachToArrayAnyRector::class);
+};
